# Unroll RowNumber Next()