@mdisibio 2023-05-01
Great idea and benchmarks!   What do you think about using `fallthrough`?   Some of the loops as-is are perfect candidates.  I think part of what's making the unrolled version quite hard to read is the two loops were combined, even though the second loop was only dependent on `definitionLevel` it is now within the switches of repetition level.  If the loops were unrolled separately and used `fallthrough` I think it could be just as fast and still quite understandable.