@joe-elliott 2023-05-01
Trying something like:
```
	switch repetitionLevel {
	case 0:
		t[1] = 0
		fallthrough
	case 1:
		t[2] = 0
		fallthrough
	case 2:
		t[3] = 0
		fallthrough
	case 3:
		t[4] = 0
		fallthrough
	case 4:
		t[5] = 0
	}

	switch definitionLevel {
	case 0:
		t[1] = -1
		fallthrough
	case 1:
		t[2] = -1
		fallthrough
	case 2:
		t[3] = -1
		fallthrough
	case 3:
		t[4] = -1
		fallthrough
	case 4:
		t[5] = -1
	}
```
which is indeed more readable, but quite a bit slower
```
name    old time/op  new time/op  delta
Next-8   131ns ± 3%   152ns ± 0%  +15.98%  (p=0.000 n=10+9)
```
and even a bit slower compared to original:
```
name    old time/op  new time/op  delta
Next-8   147ns ± 1%   152ns ± 0%  +3.26%  (p=0.000 n=10+9)
```