@joe-elliott 2023-05-01
Per an internal convo we kept the fully unrolled version but added the original implementation as `nextSlow()` to test against.

cc @mdisibio